unified mobilenet directed test components
